feat: add `libuv` support

+++ src/BaseProxyHandler.h
@@ -14,7 +14,7 @@ class BaseProxyHandler : public CProxyHandler
          * @param buffer - the buffer that we receive from upstream ( source dbs )
          * @param length - length of the buffer
          */
-        std::string HandleUpstreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context)
+        std::string HandleUpstreamData(void *buffer, ssize_t buffer_length, EXECUTION_CONTEXT *exec_context)
         {
             std::cout << "Received a Client packet..................... " << endl;
             std::cout << "Length of Packet is " << buffer_length << endl;
@@ -30,7 +30,7 @@ class BaseProxyHandler : public CProxyHandler
          * @param buffer - the buffer / response that we receive from downstream ( target dbs )
          * @param length - length of the buffer
          */
-        std::string HandleDownStreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context)
+        std::string HandleDownStreamData(void *buffer, ssize_t buffer_length, EXECUTION_CONTEXT *exec_context)
         {
             std::cout << "Received a Server packet..................... " << endl;
             std::cout << "Length of Packet is " << buffer_length << endl;
